+++ src/ops/maths.h
@@ -9,6 +9,7 @@ extern const tele_op_t op_MUL;
 extern const tele_op_t op_DIV;
 extern const tele_op_t op_MOD;
 extern const tele_op_t op_RAND;
+extern const tele_op_t op_OCT;
 extern const tele_op_t op_RND;
 extern const tele_op_t op_RRAND;
 extern const tele_op_t op_RRND;